requirements.txt
streamlit
PyPDF2
python-dotenv
langchain
openai
faiss-cpu
altair
tiktoken
huggingface-hub
InstructorEmbedding
sentence-transformers
-e .